Implemented basic TCP support layer  * buffered I/O  * design should support future expansion of protocols, proxies, etc  * so far, HTTP range requests working well (on plain HTML text tests, not yet BAM-tested)

+++ src/api/internal/io/BamFtp_p.h
@@ -38,10 +38,10 @@ class BamFtp : public IBamIODevice {
         void Close(void);
         bool IsRandomAccess(void) const;
         bool Open(const IBamIODevice::OpenMode mode);
-        size_t Read(char* data, const unsigned int numBytes);
+        int64_t Read(char* data, const unsigned int numBytes);
         bool Seek(const int64_t& position);
         int64_t Tell(void) const;
-        size_t Write(const char* data, const unsigned int numBytes);
+        int64_t Write(const char* data, const unsigned int numBytes);
 
     // internal methods
     private:
